Reimplement implicit casts. Also add support for operator casts.
